+++ public/dpk/DPKFileReader.h
@@ -34,16 +34,16 @@ public:
 	size_t				Write(const void *src, size_t count, size_t size);
 
 	// seeks pointer to position
-	int					Seek(long nOffset, EVirtStreamSeek seekType);
+	int					Seek(int nOffset, EVirtStreamSeek seekType);
 
 	// fprintf analog
 	void				Print(const char* fmt, ...);
 
 	// returns current pointer position
-	long				Tell() const;
+	int					Tell() const;
 
 	// returns memory allocated for this stream
-	long				GetSize();
+	int					GetSize();
 
 	// flushes stream from memory
 	bool				Flush();
